requirements.txt
asgiref==3.4.1
certifi==2021.10.8
Django==3.2.11
django-cors-headers==3.11.0
django-debug-toolbar==3.2.4
django-rest-framework==0.1.0
djangorestframework==3.13.1
importlib-metadata==4.10.1
pika==1.2.0
psycopg2-binary==2.9.3
pytz==2021.3
sqlparse==0.4.2
typing_extensions==4.0.1
zipp==3.7.0
